services/telemetry_input_service/requirements.txt
watchdog
kafka-python